Add assisted dialing logic to Telephony Connections.
This is a manual merge of:
I122fa1075f70095733368ed19a725f2c820a4f82

Currently, we propogate the assisted dialing property
if we detect that the client requested assisted dialing. This
is because the client is currently responsible for assisted dialing
and only passes along the request if it was successful.

Test: manual
Bug: 35963245
Change-Id: I6897554761b43449b3619747663866ff79b504ec