Enhanced permission checks for TelephonyManager#endCall() API.
The existing permission checks require CALL permission in order to end
an ongoing call or reject an incoming call.  The Telecom equivalent of
this API requires MODIFY_PHONE_STATE.  Since this API IS in fact a system
API, modifying the permission requirements to require the same
MODIFY_PHONE_STATE permission.

Test: Manual testing per bug reproduction steps.
Bug: 67862398
Change-Id: I8f15f6afd00466f6b861079a83e9fa03b962f63f