Throw on revoked location permission - telephony
When we fixed proper handling of location permisison gating
sensitive telephony calls we stopped throwing a security
exception when the permission is not held by the caller.
While this is not a security issue there is no reason to
change this behavior which is checked by CTS. This CL starts
throwing a security exception if the permission is not held.

Test: atest android.permission.cts.NoLocationPermissionTest

bug: 74074103

Merged-In: I9b275ce502d1a2bcd42405c5c713d28795086a7a
Change-Id: I9b275ce502d1a2bcd42405c5c713d28795086a7a
(cherry picked from commit 4af6628c25e52d5ae93471f56c636dad4db78c03)