Make sure manual selected network is shown in next search result.
When user want to manually select a cellular network, modem searches
available networks and return a list to NetworkSelectListPreference.
We need to make sure that if user selects a network, it will shown
as "chosen" next time user opens it again.
But challenge is, every time user opens this preference the available
networks will be different. For each network, we used to use index as
their "value" which will be later determined which was previous choice.
Now instead of relative index which changes every time, we use MCCMNC
as value since they are unique, so that however available network list
changes, it will show user's previous choice correctly.

Bug: 72029476
Test: manual
Change-Id: Ib559dbaa0da07150cfd85f395cdfd8a4399977cb