Fix issue where EmergencyDialer is slow to open in some cases.
Per recommendations by framework team, using setShowWhenLocked instead of
the window flag.
In onPause, removing the ToneGenerator teardown.  This was a source of
long pauses during the activity lifecycle.  We now initialize it only in
onCreate and tear it down in onDestroy.

Bug: 76058395
Test: Manually reproduced user switching issue; confirmed it is fixed
Merged-In: I0bd63a9c7138ab9450861293cbd438105c33eb06
Change-Id: I0bd63a9c7138ab9450861293cbd438105c33eb06