Fix the issue on reading cardId.
We should get the card from the slot first. Check whether the card is
null. If not null, get the cardId. If the card is null, just set the
cardId as null.
The index for getUiccCard() should be phoneId, which is inconsistent
with the previously given physical slot id. It's possible that
UiccController#getUiccCard() returns a null card and thus cause NPE.

Bug: 77607791
Test: test on phone
Change-Id: I6e145612398e798be0b370ccb59004c5df25f8c4