Restore the OEM hook implementation and usage
The current plan is to keep the OEM hook implementation in the 
framework, this is for legacy device and old device upgrading 
(per b/78098059).

For P launching device, the OEM hook service is removed in hal, 
so the OEM hook api in framework should return something dummy, and 
prevent infinite loop to get service, but these parts should be done
in seperate CLs after this restore CL is merged.

This reverts commit f120147b56cf529c18fc6f0de0ea8db356c4102a.

Bug: 34344851
Test: pass Treehugger tests; pass cuttlefish test RilE2eTests 
Change-Id: I8310e985ca805fcda519f5aac2b7c767aac1f4c6