Fix issue that the carrier configs are not updated after SIM unlocked
MCC/MNC are not acquired in the state of SIM lock but ICCID can be
acquired by #529328. So, the carrier config values are cached to cache
file unexpectedly at the time of fetching from CarrierService in the
state of SIM lock. As a result, cache file created with wrong values is
used after SIM unlocked and UE never refers to MCC/MNC filtered carrier
config values.

To resolve this issue, skip making of the cache file if SIM is not
loaded.

Test: manual - Checked that the carrier configs are updated after SIM
unlocked.
Bug: 113788731

Change-Id: Ie55063ce95cc1d8e1d9ff67dbdb3522dc6407de0