Update patch set 3
Patch Set 3:

(1 comment)

> Do we need to care about the other places at this point though? I don't care as much if users get some broken functionality elsewhere that is less critical than emergency calls.

Maybe we don't, but I haven't looked enough to know this for sure, the code is really spread out.

We could probably just check if Google Dialer can do emergency calls when it's not a system app and not the default dialer, both when the device is locked and unlocked.

Maybe there's really no difference between system dialer apps and third party dialer apps when it comes to emergency calls (I'm talking about stuff such as enforceCallingOrSelfPermission() where some system privilege is needed) and this is just to ensure that a previously tested dialer is used.

Patch-set: 3