Restore the OEM hook implementation and usage
The current plan is to keep the OEM hook implementation in the 
framework, this is for legacy device and old device upgrading 
(per b/78098059).

For P launching device, the OEM hook service is removed in hal, 
so the OEM hook api in framework should return something dummy, and 
prevent infinite loop to get service, but these parts should be done
in seperate CLs after this restore CL is merged.

This reverts commit baafa294d1b1fc4067b334851ec62564537bc6fa.

Bug: 34344851
Change-Id: I6e49d2b34c5f34f59d7cae1fc20b3881ee677642
Test: pass Treehugger tests; pass cuttlefish test RilE2eTests

Mergde-In: I8310e985ca805fcda519f5aac2b7c767aac1f4c6