Ensure CDMA emergency tone/vibration respects ringer mode.
Where the CDMA emergency setting is set to "tone" mode, suppress the tone
when the ringer mode is VIBRATE or SILENT.
Where the CDMA emergency setting is set to "vibrate" mode, suppress the
vibration when the ringer mode is SILENT.

Test: Set emergency tone mode to "tone".  Ensure tone is heard when phone
is in normal mode.
Test: Set emergency tone more to "vibrate".  Ensure vibration happens
when phone is in normal or vibrate mode.
Bug: 79122864

Change-Id: I8f3b844f0f438e815201562991737d068d463aa7