Hide Wi-Fi and Video calling preferences if IMS service is not ready
CallFeatureSettings checks IMS service state and hides Wi-Fi and Video
calling preferences if IMS service is not ready. MobileNetworkSettings
has the same setting but it is not checked on it. MobileNetworkSettings
should also check IMS service state and hide them.

Bug: 109916999
Test: manual - Check that Wi-Fi and Video calling preferences
are hidden when the IMS service state is not ready.
Change-Id: Icee297c562b5985d3923dbbe0ccca22ba5e88e26