Update patch set 1
Patch Set 1:

It writes it to the SIM, so it's persistent through wipes. We have had this feature since before gingerbread, but for some reason it didn't make it into 14.1 and 15.1

Normally stock roms doesn't have this option, so the only way one could program the number was doing it via an iPhone.

> hmmm, on my 2nd attempt to set the phone number it gets stuck on updating settings.

A dual-sim issue maybe?

Patch-set: 1
Reviewer: Arne Coucheron <444@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>