Fix wrong call forwarding preferences are displayed in multi-window mode
Call forwarding preferences become gray out and summary text of enabled
preferences back to default text. Because the enable state of call
forwarding preferences are not saved and restored.

To resolve the issue, save and restore call forwarding preference enable
state when changing to multi-window mode.

Bug: 111378366
Test: manual - Checked call forwarding preferences are displayed the
same as before when changing to multi-window mode.

Change-Id: I4ef0f119ddf231bc237d6292eab92d5d6d945cd0