Fix issue that preferences of call settings cannot be disabled
Video calling and Wi-fi calling preferences cannot be disabled when
SIM2 receives an incoming call while call settings for SIM1 is
displayed. The cause is that it registers a phone listener only for
a subId corresponding to currently displayed call setting.

To resolve this issue, listen phone state by the default
TelephonyManager to receive phone status of all subs.

Test: Checked that Video calling and Wi-fi calling preferences are
disabled when SIM2 receives an incoming call while the call
settings for SIM1 is displayed.
Bug: 74414829

Change-Id: Id97965e93f9fc6727164dc9de64f51bc014ab06d