Fix the issue that ims emergency call redial fails
Ims emergency call redial fails because it fails to switch a phone from
failed phone to other phone to retry the emergency call.
Removing ims phone from a list of emergency retry phones fails because
the list consists of default phones and ims phone is not in the list.

Bug: 113303907
Test: manual - Checked that when an ims emergency call failed it will
be redialed on another phone
Test: auto - Passed TelephonyConnectionServiceTest
Change-Id: I24096c0b27822de2cc626e2e1e01029febc32458