Update enabled networks entries on rotation
The save/restore logic in ListPreference/DialogPreference doesn't
preserve changes to the entries, so we have to manually update the
entries after rotation.

Currently on rotation we update the entries in mButtonEnabledNetworks
dialog too late. Since the dialog is created in onRestoreInstanceState,
we can make sure it's correct by updating the dialog in onCreate instead
of onActivityCreated.

Bug: 112343168
Test: manual
Change-Id: Ib7527e48d79393a82e6236c34c1fe0a226ce717e