Handle carrier config changes in TelecomAccountRegistry.
TelecomAccountRegistry makes use of the carrier configuration to set a
number of properties on the Telephony phone account(s).  This is done when
there is a subscription change.
It is, however, possible for a carrier config change to occur outside of
a subscription change.  If this happens, the change will not be reflected
in the accounts registered by the TelecomAccountRegistry.
Added a receiver for ACTION_CARRIER_CONFIG_CHANGED which triggers a
rebuild of the telephony phone account.  To ensure that unnecessary updates
are not made to the phone account in Telecom, compare the new to the old
phone account definition in this case.  If there is no change, we do not
update the phone account.

Test: Manual testing -- verify that carrier config change causes update to
the phone account for Telephony.
Bug: 112884689

Change-Id: Ia45023251583ea15bfacd85253c88bd79fe1dc27