Remove unused code-paths in Telephony.
Removed some methods from PhoneInterfaceManager; these APIs were
deprecated in previous releases, and are now @remove.  The code for these
is replaced by code in Telecom, so they're no longer required.

Elsewhere, there is a lot of code in other classes which can be removed
now that those APIs are removed.

CallLogger is not used at all, so can be removed.

CallNotifier had some audio state change code which is in the
InCallTonePlayer subclass.  Most of that class is now in the Telecom
equivalent.  However, the disconnect tones are now handled in
Telecom, so removed an audio state reset which takes place in there.

CallGatewayManager is almost entirely not used; calls were being added
to the connection map, but the data in there was not being consumed
anywhere.  Kept the class around because EMPTY_INFO is used in the
PhoneUtils placeCall method (this is still needed for OTASP unfortunately).

Test: Compile
Bug: 115776310
Change-Id: Id599d093b8c21213c74d5e95f80d9f1082c9de43