Apply new network selection api on NetworkSelect fragment
This applies the new manual network selection api on NetworkSelect
fragment.

We also notifies the NotificationMgr when the network selection is
changed by the user, so the Network selection fragment no need to call
NotificationMgr.updateNetworkSelection when the network selection is
failed.

Bug: 110820484
Test: manually test
Merged-In: I8bc50d270edd9bbd556bad235122b0bfe366222e
Change-Id: I8bc50d270edd9bbd556bad235122b0bfe366222e