Prevent turning on airplane mode when in network identified e-call.
When the device is in a network-identified emergency call, prevent the
user from toggling on airplane mode.  Instead of relying on old
PhoneUtils code to see if there is an emergency call, rely on Telecom
which knows of all calls.

Test: Used new test intent to identify call as ecall and verified user is
not able to turn on airplane mode.
Test: Set outgoing number as emergency number using ril.ecclist property
and verify user is not able to turn on airplane mode.
Bug: 77565333

Change-Id: Idb8d4745a1651097c7276233bc06cc8cb254ac96