Treat single party IMS conference as a standalone call.
Where an IMS conference drops to a single participant, remove that child
from the conference and tell telecom to no longer treat the conference as
if it is a conference.  Also pass through the participant name/number to
telecom.  When recalculating whether manage conference is supported, do
not allow manage conference if there is just a single participant.

Test: manual
Bug: 75975913
Change-Id: Iba9a65bc37df5df6472c51183410b774219c9eb2