Add Phone Object to MainThreadRequest
Add support for sending MainThreadRequests using
a Phone object instead of a subId. This is the first
step to permitting other invocation patterns from
PhoneStateListener, and it facilitates cleaner code
in the one case where ITelephony already operates on
a specific phone rather than the default phone, a
slot or a subId.

Bug: 26569588
Test: SL4A - getAllCellInfo() (targetSdk=27)
Change-Id: I72905cf3c26debdfdd540513b1345d23be851eda