Revert "Fix the duplicated info of SIM cards after Hot-plug." This reverts commit caca9048f86c5d954277f5bedd781fae604ae4e0.
The CL being reverted causes the Calling Accounts screen to crash due
to a null pointer exception.
The issue appears related to the fact that subscription manager
change callback fires after the state of the preferences on the screen
has already been impacted in onResume.
Maybe this works for multi-sim devices, but on single sim devices this
causes a crash on this preference screen all the time.

Reason for revert: CL causes a crash when impacted screen is opened.
Test: Manual
Bug: 117541504

Change-Id: I557333712080c050f036d6eb1752295ad4cafbf1