Fix multisim issue with sim call manager.
Ensure that on multisim devices we only configure the sim call manager
when its the default sim manager for the current sub.
I considered making getSimCallManager multisim aware, but it works out to
be a pretty large change for a scenario that I don't think will reasonably
happen very often.  Ultimately we would like to deprecate the sim
call manager in the future.

Test: Manual
Bug: 111411739
Change-Id: I6c9e7f30a249e11669108fbb4175ecbf192d80d5