Add UT capability detection and caching
1) add the ability to set/query IMS provisioning
status for MMTEL capabilities.
2) Add sharedprefs cache for UT provisioning, since
it does not use an ImsConfig value internally in
the modem. This cache will be used by ImsManager
to determine whether to enable/disable the UT MMTEL
capability when the carrier config changes. Whether
or not we use the cache is determined by the new
ut provisioning required carrier config.
3) Do not throw IllegalArgumentException if the
subscription no longer is active. Instead, just
return error codes for Provisioning.

Bug: 119896853
Test: atest FrameworksTelephonyTests
Change-Id: If99b459dd4f08bde0059b2fd14a2a2c132cadf30