Enable shortcut view only if it's launched from power menu.
The purpose of shortcut view to provide a easy way for users to make an
emergency call. However, easy-to-access also means easy-to-be-mis-
triggered.

The chance to mis-dial an emergency number from power menu should be
much lesser than launching from other entry points.

In addition, we need more log for analyzing mis-dial bugs.

Bug: 112168722
Bug: 124419284
Test: Manually
Change-Id: I3bd3ae08881c6bb81ee4613b8d1cea3b88fc934a