Fix issue where device does not create any phone accounts.
A code change was recently made to TelecomAccountRegistry which made it
possible for the emergency-only phone account to NOT be added.
Hard-wiring setupAccounts in this manner is not correct as the method has
a check at the end to see if the dfeault account should be added.

I've corrected the method short-circuit potentials and also move the
creation of the default emergency-only phone account into a finally
block to ensure coding errors (or crashes for that matter) in the main
loop will not cause a situation where there is no emergency call acct.


Test: Removed sim, rebooted and noted that the default account is created.
Test: Removed and replaced sim numerous times and noted that the correct
default emergency and regular phone accounts are created.
Bug: 126214468

Merged-In: Iaef69a5331078a9f848cc00e7e98e46b3a8862a6
Change-Id: Iaef69a5331078a9f848cc00e7e98e46b3a8862a6