Consider subId when loading Call Settings and Dialer status hint strings
Ensure that the correct string resources is loaded in dual sim scenario
in screen Call Settings and in Dialer status hint by considering subId.

This is done by retrieving the Resources with a call to
SubscriptionManager.getResourcesForSubId(). This ensures that the
correct string resources is loaded if an operator has customized any of
these strings.

Bug: 144051856
Test: atest TeleServiceTests
Change-Id: Ie1b1dada1b1181c0030435a3bf5f95500db5f61d