Fix wrong wording of a notification and dialog in IMS ECM
Mobile data is not allowed during CDMA ECM. So, the device notifies the
restriction with a notification and dialog like "Data connection
disabled". But there is no such restriction in IMS ECM. However, the
same wording as CDMA ECM is used in IMS ECM.

There are three UI to display wrong wording.
1. Notification during ECM
2. Confirmation dialog to exit ECM
3. Notification during an emergency call

To resolve the issue.
Fix the wording for IMS ECM.

Test: manual - Checked correct wording is displayed in the notification
and dialog.
Bug: 141287026

Change-Id: Ibdda36993d974fba15aba249d52c7a83909d8b6a