overrideConfig should use existing notification flow for CCC
When CarrierConfigManager#overrideConfig was called, we were calling
the CCC broadcast directly, instead of using the existing
pattern of calling notifySubscriptionInfoUpdater. This may become
an issue if future CTS/GTS tests depend on this happening.

Also, only allow the SHELL UID to call overrideConfig, as
it should only be used for testing.

Bug: 130372397
Test: Manual
Change-Id: I945248d325b7cc662b9ae6dac99995fd6b8ef6ec