Trampoline SmsManager SimDialogActivity intent
Start the SimDialogActivity using a transparent
"trampoline" activity in order to process the
activity's result in startActivityForResult.

1) Create PickSmsSubscripionActivity, whose only
purpose is to start SimDialogActivity and process
the result. This is called by SmsManager to process
operations on MSIM devices, where the user has
chosen "ask every time" as their default SMS
subscription.

2) Create SmsManagerTestApp app, which sends an SMS
using a foreground activity as well as a background
service for testing.

Bug: 130853716
Test: run SmsManagerTestApp
Change-Id: I44b2c352548e5b61f61250547bfd3261b5060e5a