Mark DDS PhoneAccount as CAPABILITY_EMERGENCY_PREFERRED
On MSIM devices where GNSS SUPL is not supported for
emergency calls on the non-DDS subscription, mark the
DDS subscription as CAPABILITY_EMERGENCY_PREFERRED so
that Telecom will override the user's choice of
PhoneAccount and always try to place the emergency call
over the DDS subscription.

Note: We do not set this preference if the subscription is
marked as opportunistic. If the DDS is a data only SIM, we
will dial on the non-DDS when Telephony receives the
outgoing call request.

Bug: 131203278
Test: Manual Testing; atest FrameworkTelephonyTests
Merged-In: Ia01416069eeceb3a8e01b7e2aa0121b2ab43289c
Change-Id: I6beb1955bb0aa4140e1ce5b17d51d249c4b85106