Disable single party call emulation when a connection manager is in use.
Where a conference is hosted on a carrier which uses a connection manager,
we need to disable the single party call emulation when there is a single
participant in the conference.  This is due to the fact that there are some
signals which must be sent to Telecom which is not possible to sent
reliably to Telecom. This includes whether the conference is being treated
as a conference or not, the conference connect times, and the address
associated.  The connection manager wraps a RemoteConference and provides
a Conference instance back to Telecom.  It is not, however, possible for
the framework to know which RemoteConference corresponds to which
Conference.  As a consequence of this, we cannot reliably pass the signals
required for single party call emulation back to Telecom.

To solve this bug, we disable single party call emulation where there
is a connection manager in use.

Test: Manual test on FI; verify issue no longer reproduces and that single
party call emulation is not present.
Test: Manual test on Verizon; verify single party call emulation works.
Test: Added ImsConference unit test.
Bug: 127305753

Change-Id: I0bcbd46f9ae285b154f8819402a55c4a994edaa6