Fix where post-disconnect CEP causes single party call mode to exit.
On some carriers, a conference event package will be sent just prior to
disconnecting the conference.  The existing logic in ImsConference would
interpret this as no longer being a single party call and enter conference
mode again, but with no participants.  As a consequence the call would not
be logged.

Test: Wrote failing unit test and fixed code to verify that the problem
is fixed.
Bug: 133323379

Change-Id: I092c0f06ce26d3bafdca2173083de2c36acdc037