Do not pass EVENT_MERGE_START for ImsConference.
A change was made to pass other connection events from the host connection
to the ImsConference.  This means that the host conference will pass
EVENT_MERGE_START for conferences, which was not done in the past (it is
normally only sent for calls being merged INTO a conference, not on the
conference itself).

Test: Manually reproduce and verify that the merge start event is not
passed.
Bug: 133128377

Change-Id: Ib1d28909f3d78416b3cebd21848b033a8b88e261