Handle disconnected participants which remain in CEP data.
Some carriers will report a disconnected participant in the CEP but then
never remove it from a subsequent CEP update.
This can render the "single participant emulation" inactive as the device
never thinks there is just a single participant in the conference.
This change updates the conference event package handling to check for
disconnected participants and ensure they're cleaned up appropriately.

Test: Add new unit tests for the bug scenario; verify they pass.
Bug: 111860217
Change-Id: I3fd8271368b0023badcfaec3f25564971fed5d06