Remove 110 for German for legal requirement.
Remove 110 for country code "de" from framework database.

In Germany the federal presidents decide whether 110 and 112 is routed
to emergency centre or 110 is routed to police department and 112 to
fire department. The operators needs to make sure it is routed as
required.

Bug: 137052208
Test: Hard code network country code to "de". Only 112 shows up in
Faster Emergency Launcher.

Change-Id: Ib5385fbdff736cc5795c9bbdab0e591a83b312bd
Merged-In: Ib5385fbdff736cc5795c9bbdab0e591a83b312bd