RTT changes for ignoring user-set setting
Check the carrier config for the ignore user-set RTT mode flag. If it's
on, don't propogate the user-set RTT mode setting to the subscription
that it's on for, and always return from isRttEnabled as if the user-set
RTT mode setting were on.

Bug: 136035164
Test: manual
Change-Id: I8039ba8061b83e315af7fb7007253978f9544100
Merged-In: I8039ba8061b83e315af7fb7007253978f9544100