config files has matching carrier id which passed to carrier Services.
this is to fix a race condition triggered from a back-to-back
sim refresh events. scenario is when we passed carrier id to
carrierconfigservices, sim status change to absent and thus we
passed unknown carrier id. later carrierconfigservices returns the
empty result for unknown carrier id, sim status change to loaded
and we persist the empty config to a cache file with correct
carrier id. The fix is make sure config file always has matching
carrier id as what passed to carrierconfig services.

Bug: 137951167
Test: Manual
Change-Id: I9960d167005b3e23d30692074d877eca08606f7d