Post Connection Events to handler.
The disconnect messages are posted to the handler, but the connection
events were not.  This meant when merging calls into a conference, the
event to indicate the calls merged successfully would come BEFORE the
disconnect signals arrived in Telecom.
Putting the connection events to the handler ensures they arrive AFTER
the disconnect signals.

Test: Described in other CL on topic.
Bug: 139169804
Change-Id: I7491504032538e8ccc87cfc65b5eaaeb31e32de0
Merged-In: I7491504032538e8ccc87cfc65b5eaaeb31e32de0