DUT fail to unlock SIM after selecting Home IMSI from STK
When unlock SIM pin,SIM begin refresh,and the UiccCardApp was disposed.
But the unlock process was canceled.
Therefore, the previous unlocking process was interrupted without returning,
and the keyguard PIN lock View was always displayed.

If UiccCardApp was disposed,send the CommandException.Error.ABORTED as result for the previous Message.

Test: manual

Bug:140907125

Change-Id: I941dc5ce8bc8ea903ddbe6b5fa88292636538107