Remove dependence on Connection.Listener in Telephony.
Telephony used the Connection.Listener class a LOT, even though its an
@hide detail of the Telecom Connection API.

Summary of changes:
In TelephonyConnection, created a parallel to the Connection.Listener
for the events which telephony internally cares about. This required making
a lot of TelephonyConnection versions of the set methods so we could
trigger the right listeners.
In ImsConference, removed dependence on listener to track when a connection
is destroyed; the destroy request originates in this class so we don't
really need to go through the listener indirection.


Bug: 141576016
Test: Manual smoke test.
Test: Run unit tests.
Test: Run CTS tests.
Change-Id: I037c59a0aa75c2f4885a20642ec95c750175fe21