IMS: Recalculate conference on RTT mode change
When there is RTT mode change there will not be any call state change triggered
to UI as there is no change in call state or call details but only the call
capability is updated. This gives incorrect data to InCallUI on confernceable
calls. Recalculate confernce on RTT mode change to update UI on conferenceable
calls.

Test: Manual

Bug: 135572135
Change-Id: I976719fcd3e4ee29765320b527c96468a2f16d76