Call Forward Setting preference is disabled when setting an empty number
1.Launch phone app-> Press menu key-> Settings-> Call forwarding.
2.Select Always forward-> empty the number in the edit box
3.Click "enable"

You will see an error dialog. After the dialog is dismissed,
the "Always Forward" preference is disabled and not
interactive until the setting menu is rei-entered.
The same issue exists for other CF preference in
the similar case.

[Root Cause]
When the CF is updated with an empty number, lower layer will
return an error and the preference will be disabled.

[Resolution]
Should not disable the preference when the response failed.
End user can have a chance to update the number and redo it.

Fixes: 145515029
Test: Manual
Change-Id: I27a83075658ca164d2e5450862c5a9ebed58cc34