Split TeleService into a client apk and a lib.
The client apk has the manifest and the lib has the code
and resources. This is being done so that the apk can run
under the phone shareduid, while the lib can be part of telephony
mainline module and be updatable.

Test: basic telephony sanity & atest TeleServiceTests
Bug: 144806641
Merged-in: I3cd7dd692ec9325a10aa7ae63e77240c7ad58112
Change-Id: I3cd7dd692ec9325a10aa7ae63e77240c7ad58112