Add support to fetch carrier config for no SIM case.
When there is no SIM card, CarrierConfigManager returns the code
default values to the clients. This is a problem for certain carriers
if they wanted a different behavior for no SIM case.

This change queries the default carrier config app with null
carrier id when the SIM card is absent. This will help OEMs
to override the code default values for no SIM cases.

Bug: 142090499
Test: Manually verified by removing SIM card
      Manually verified by rebooting w/o SIM card

Change-Id: I4f4c0b7716d298c0fd43c0a074b16465c6944ea5