Set RTT capability on the emergency PhoneAccount
When a device doesn't have a SIM inserted, Telephony registers a special
PhoneAccount with Telecom for making emergency calls. This changes sets
the RTT capability on that emergency PhoneAccount under certain
circumstances, letting the phone dial an emergency RTT call even without
a SIM.

In order for a device to support this functionality, they should set the
config_support_simless_emergency_rtt config to true in their device
overlay.

Fixes: 147078534
Test: manual
Change-Id: Idab399e99f4a25396003329e6d03048d8767ba7b