Fixed that OK button is not localized
Because the dialog is stored as a static variable, so
the ok button string resource was not refreshed after
language changed.

And in fact there is no need to cache the dialog. We
should always create it when needed, and let it GC'd
when closed.

Test: Manual
Fix: 145309447
Merged-In: Iaccf51e669dbf76ae4c9fcf149cec665f9090c97
Change-Id: Iaccf51e669dbf76ae4c9fcf149cec665f9090c97
(cherry picked from commit 07666003e20dc99b50f11ddb540d9b2fdf3092ec)