Support local disconnect of empty IMS Conference.
Add support for locally disconnecting an IMS conference when the
participant count goes to 0.
Refactor carrier config access in ImsConference to use a builder passed
in, facilitating easier testing and abstraction of carrier config from
the ImsConference class.

Test: Add unit test to verify local disconnect when participant count is
zero.
Test: Use conference event package test cmd/intents to inject test CEP to
simulate a conference dropped to 0 participants.  Verify IMS conference is
disconnected.
Fixes: 151707520

Change-Id: I024b6e8705640460c88d8418256b59beaa90362f