Add unit test for CEP at conference creation.
There is a case where a CEP can be received BEFORE a conference is created.
Adding a test to ensure that if a single participant is present in the
CEP data at the start of an ImsConference that we will correctly set
the conference property to indicate it is not a conference.

Test: This IS a test.
Fixes: 134047008
Change-Id: Id563d8caac6175ee569cffa5cb26ea26b6b4bdc7