Fix comparison of SIM State values.
A recent change was made for sim personalization; the comparison of
bag.mIccStatus was using the == operator, when realistically it should use
the .equals() method. mIccStatus originates from an intent's extras
so we cannot guarantee that it is the exact same instance of the constant.
However, using .equals ensures we are comparing the actual value, which is
what we did prior to the sim personalization changes.  As a result the
PUK progress dialog would NOT be hidden when the sim became ready.

Also, added some more log statements for the show/hide of this activity
just in case this was a red-herring and was not REALLY the cause of the
bug being seen.

Test: Manual
Fixes: 154672821
Change-Id: I66063d09cf38d223e8dac882521cf610715b1e23