Fix ArrayIndexOutOfBoundsException in IccNetworkDepersonalizationPanel
The arrays in the class IccNetworkDepersonalizationPanel were created with the length by calling the api "TelephonyManager#getSimCount".
When the active modems count was changed, the length of the arrays sShowingDialog and sNdpPanel had not been updated.

Initialized the array length by calling the API "TelephonyManager#getSupportedModemCount" instead.

Bug: 158014107
Test: manual
Change-Id: Id31ac22a1581daf68769c967494dab05c02983db
Merged-In: Id31ac22a1581daf68769c967494dab05c02983db