Add an exported flag in manifest
With b/150232615, we will need an explicit value set for the exported
flag when intent filters are present, as the default behavior is
changing for future versions. This change adds the value reflecting the
previous default to the manifest.

These changes were made using an automated tool, the xml file may be
reformatted slightly creating a larger diff. The only "real" change is
the addition of "android:exported" to activities, services, and
receivers that have one or more intent-filters.

Bug: 150232615
Bug: 159134989
Test: TH
Exempt-From-Owner-Approval: mechanical refactoring
Change-Id: Ib218fd921b963ed8aa32c8f74581c9b385d60a9e
Merged-In: Ib218fd921b963ed8aa32c8f74581c9b385d60a9e