Handle potential SecurityException in PstnIncomingCallNotifier.
There is a potential for the incoming call phone account to not exist
in some rare cases.  I suspect this involves a profile change on
MVNO carriers which aggregate other carriers together.  This change
catches the exception Telecom throws when the phone account is no longer
registered and logs a WTF so that crash clustering can be performed on
the issue.  The logs include detailed information about the phone account
state on the device so we can analyze this further.  The incoming call is
rejected instead of crashing the device.

Test: Change code so that a security exception is always thrown; verify
in logcat that the crash logs include detailed information about the
call and phone account state at the time of the problem. Verified that
the call is silently rejected in this case.
Fixes: 147978774

Change-Id: I7a000a70b3a8b8883d4aaa8d34c99e4eb33335bb