Handle P-Associated-Uris which are SIP uris in conference host check.
When checking whether a CEP participant is the host device, we previously
assumed the P-Associated-Uri entries were just tel URIs.  Although in
practice this has worked so far, there are some carriers where the mix of
P-Associated-Uri entries is JUST a list of SIP URIs, with none being
a TEL URI.

This change assumes the P-Associated-Uri can also be SIP URIs and adds
tests to confirm behavior in the existing and newly identified cases.

Test: Added unit tests.
Fixes: 172290149
Change-Id: I83df998c655ab61246432e846b76b648a5526d79