Stop mocking IRoleManager.
IRoleManager isn't an API and is going to be moved into module, so it
won't be available anymore. We could have used mockito-inline to mock
the final RoleManager, however this test is instrumenting against
com.android.phone which can't be made debuggable. So introduce an
adapter class similar to what SipTransportController is already doing.

Bug: 158736025
Test: atest RcsProvisioningMonitorTest
Change-Id: I1b9c8d86f459a2391680291d07a70bda614a608d